exercises/exercise-8/recommender_model/requirements.txt
scikit-surprise==1.1.4
numpy==1.26.4
pandas
flask
requests